renamed ks_test directory to kstest
