+++ explainability-core/src/main/java/org/kie/trustyai/metrics/drift/kstest/GKSketch.java
@@ -1,4 +1,4 @@
-package org.kie.trustyai.metrics.drift.ks_test;
+package org.kie.trustyai.metrics.drift.kstest;
 
 import java.util.ArrayList;
 import java.util.List;
