+++ explainability-service/src/main/java/org/kie/trustyai/service/endpoints/metrics/drift/ApproxKSTestEndpoint.java
@@ -11,8 +11,8 @@ import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.jboss.logging.Logger;
 import org.kie.trustyai.explainability.model.Dataframe;
 import org.kie.trustyai.metrics.drift.HypothesisTestResult;
-import org.kie.trustyai.metrics.drift.ks_test.ApproxKSFitting;
-import org.kie.trustyai.metrics.drift.ks_test.ApproxKSTest;
+import org.kie.trustyai.metrics.drift.kstest.ApproxKSFitting;
+import org.kie.trustyai.metrics.drift.kstest.ApproxKSTest;
 import org.kie.trustyai.service.data.cache.MetricCalculationCacheKeyGen;
 import org.kie.trustyai.service.payloads.metrics.BaseMetricRequest;
 import org.kie.trustyai.service.payloads.metrics.MetricThreshold;
