+++ explainability-service/src/main/java/org/kie/trustyai/service/endpoints/metrics/drift/KSTestEndpoint.java
@@ -11,7 +11,7 @@ import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.jboss.logging.Logger;
 import org.kie.trustyai.explainability.model.Dataframe;
 import org.kie.trustyai.metrics.drift.HypothesisTestResult;
-import org.kie.trustyai.metrics.drift.ks_test.KSTest;
+import org.kie.trustyai.metrics.drift.kstest.KSTest;
 import org.kie.trustyai.service.data.cache.MetricCalculationCacheKeyGen;
 import org.kie.trustyai.service.payloads.metrics.BaseMetricRequest;
 import org.kie.trustyai.service.payloads.metrics.MetricThreshold;
