+++ explainability-service/src/test/java/org/kie/trustyai/service/endpoints/metrics/drift/ApproxKSTestEndpointTest.java
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.kie.trustyai.explainability.model.Dataframe;
-import org.kie.trustyai.metrics.drift.ks_test.ApproxKSFitting;
-import org.kie.trustyai.metrics.drift.ks_test.ApproxKSTest;
+import org.kie.trustyai.metrics.drift.kstest.ApproxKSFitting;
+import org.kie.trustyai.metrics.drift.kstest.ApproxKSTest;
 import org.kie.trustyai.service.endpoints.metrics.MetricsEndpointTestProfile;
 import org.kie.trustyai.service.mocks.MockDatasource;
 import org.kie.trustyai.service.mocks.MockMemoryStorage;
